+++ database/migrations/2020_02_09_024408_create_field.php
@@ -15,9 +15,9 @@ class CreateField extends Migration
     {
         Schema::create('field', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('broker_id')->unsigned();
-            $table->integer('city_id')->unsigned();
-            $table->integer('area_id')->unsigned();
+            $table->bigInteger('broker_id')->unsigned();
+            $table->bigInteger('city_id')->unsigned();
+            $table->bigInteger('area_id')->unsigned();
             $table->foreign('broker_id')->references('id')->on('broker');
             $table->foreign('city_id')->references('id')->on('city');
             $table->foreign('area_id')->references('id')->on('area');
